can now delete cart products when not logged in

+++ resources/js/components/Cart/Cart.js
@@ -8,13 +8,23 @@ export default {
 			livraison:25,
 			finalPrice:0,
 			productToDelete: '',
+			id:document.querySelector("meta[name='user-id']")
 		}
 	},
 	methods:{
 		deleteProduct : function (event) {
-			axios.delete('cartItem/' + event.id).catch(error => {
-				console.dir(error);
-			})
+			if (this.id != null) {
+				console.log(event.id)
+				axios.delete('cartItem/' + event.id).catch(error => {
+					console.dir(error);
+				})
+			}else{
+				var test = JSON.parse(localStorage.getItem('storedID'));
+				var removeIndex = test.map(function(item) { return item.id; }).indexOf(event.id);
+				test.splice(removeIndex,1);
+				localStorage.setItem('storedID', JSON.stringify(test));
+				this.products = JSON.parse(localStorage.getItem('storedID'));
+			}
 		}
 	},
 
@@ -33,7 +43,7 @@ export default {
 		let id = document.querySelector("meta[name='user-id']")
 		if (id != null) {
 			this.products = JSON.parse(this.cart);
-						console.log(this.products)
+			console.log(this.products)
 
 
 		} if(id == null) {
@@ -41,7 +51,6 @@ export default {
 		}
 		var finalsubPrice= 0;
 		this.products.forEach(function(product) {
-			console.log(product.quantity)
 			var total =  product.price*product.quantity;
 			product.totalprice=total
 			finalsubPrice=finalsubPrice+product.totalprice;
